feat: webserv.hppにenum STATUS_CODEを移動
