+++ includes/webserv.hpp
@@ -8,4 +8,18 @@
 #define BUFFER_SIZE 1024
 #endif
 
+namespace http_status_code {
+enum STATUS_CODE {
+	OK = 200,
+	CREATED = 201,
+	NO_CONTENT = 204,
+	PERMANENT_REDIRECT = 308,
+	BAD_REQUEST = 400,
+	FORBIDDEN = 403,
+	NOT_FOUND = 404,
+	METHOD_NOT_ALLOWED = 405,
+	INTERNAL_SERVER_ERROR = 500,
+};
+} // namespace http_status_code
+
 #endif
